// management/frontend/src/api/schemas/systemUsersReadUsersParams.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Ego Collection Management
 * OpenAPI spec version: 0.1.0
 */

export type SystemUsersReadUsersParams = {
/**
 * @minimum 0
 */
skip?: number;
/**
 * @maximum 500
 */
limit?: number;
username?: string | null;
is_active?: boolean | null;
};