// management/frontend/src/api/schemas/userPublic.ts
/**
 * Generated by orval v8.6.2 🍺
 * Do not edit manually.
 * Ego Collection Management
 * OpenAPI spec version: 0.1.0
 */

export interface UserPublic {
  id?: string;
  created_at?: string | null;
  updated_at?: string | null;
  deleted_at?: string | null;
  creator_id?: string | null;
  updater_id?: string | null;
  deleter_id?: string | null;
  /**
     * @minLength 1
     * @maxLength 150
     */
  username: string;
  avatar?: string | null;
  status?: number;
  is_active?: boolean;
  is_root?: boolean;
  last_login?: string | null;
  creator_name?: string | null;
  updater_name?: string | null;
}