// management/frontend/src/api/zod/system-users/system-users.ts
export const systemUsersReadUsersQuerySkipDefault = 0;
export const systemUsersReadUsersQuerySkipMin = 0;

export const systemUsersReadUsersQueryLimitDefault = 20;
export const systemUsersReadUsersQueryLimitMax = 500;



export const SystemUsersReadUsersQueryParams = zod.object({
  "skip": zod.number().min(systemUsersReadUsersQuerySkipMin).default(systemUsersReadUsersQuerySkipDefault),
  "limit": zod.number().max(systemUsersReadUsersQueryLimitMax).default(systemUsersReadUsersQueryLimitDefault),
  "username": zod.union([zod.string(),zod.null()]).optional(),
  "is_active": zod.union([zod.boolean(),zod.null()]).optional()
})